requirements.txt
pytest
requests
ddt
selenium
PyYAML
